+++ lib/flutter_auth_provider.dart
@@ -19,11 +19,25 @@ class AuthProvider<U> extends ChangeNotifier {
   bool get isLoggedIn => _currentUser != null;
 
   /// Initialize AuthProvide with [AuthStore] implementation to store logged in [U].
-  AuthProvider(this._store) {
-    _initialize();
+  AuthProvider(this._store);
+
+  /// Initialize AuthProvide with [AuthStore] implementation to store logged in [U] with [LoginListener] and/or [LogoutListener].
+  factory AuthProvider.withListeners(
+      {required AuthStore<U> store,
+      LoginListener<U>? onLogin,
+      LogoutListener? onLogout}) {
+    AuthProvider<U> provider = AuthProvider(store);
+    if (onLogin != null) {
+      provider.addLoginListener(onLogin);
+    }
+    if (onLogout != null) {
+      provider.addLogoutListener(onLogout);
+    }
+    return provider;
   }
 
-  _initialize() async {
+  /// Call [AuthStore] retrieve to load user from memory.
+  Future<void> initialize() async {
     final user = await _store.retrieve();
     _setUser(user);
   }
@@ -39,6 +53,7 @@ class AuthProvider<U> extends ChangeNotifier {
   }
 
   /// This function should be called when you want to logged out current user.
+  /// Store will get cleared and [LogoutListener] will get called.
   logout() async {
     await _store.delete();
     _logoutListeners.forEach((listener) {
@@ -55,14 +70,14 @@ class AuthProvider<U> extends ChangeNotifier {
   /// Add [listener] to LoginListeners.
   /// The listeners will get called with logged in user [U].
   /// Setting up Sentry user and other login related activities can be done using the listener.
-  void addLoginListener(LoginListener<U> listener) {
-    _loginListeners.add(listener);
+  void addLoginListener(LoginListener<U> onLogin) {
+    _loginListeners.add(onLogin);
   }
 
   /// Add [listener] to LogoutListeners.
   /// The listeners will get called when user logout is called.
   /// Clearing Sentry user other cache clearing can be done using the listener
-  void addLogoutListener(LogoutListener listener) {
-    _logoutListeners.add(listener);
+  void addLogoutListener(LogoutListener onLogout) {
+    _logoutListeners.add(onLogout);
   }
 }